# src/main/resources/openapi/api.yaml
openapi: 3.0.1
info:
  title: Student Tutor Booking Management System
  description: This is a system allowing students and tutors to book in-person sessions
    at designated venues on campus.
  termsOfService: http://swagger.io/terms/
  contact:
    email: 222641495@mycput.ac.za
  version: 1.0.0
servers:
  - url: https://localhost:8080
...
paths:
...
  /student:
    get:
      tags:
        - student
      summary: Get all student
      description: Returns all student in the system
      operationId: getAllStudents
      responses:
        '200':
          description: Retrieved students successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudentDTO'
        '400':
          description: Error retrieving students
          content: { }
    post:
      tags:
        - student
      summary: Add a new student
      description: This can only be done by the admin.
      operationId: addStudent
      requestBody:
        description: StudentDTO object that needs to be added to the system
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentDTO'
        required: true
      responses:
        "201":
          description: StudentDTO added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentDTO'
        "400":
          description: Invalid input
          content: {}
      x-codegen-request-body-name: body
  /student/{studentId}:
    get:
      tags:
        - student
      summary: Get student by ID
      description: Returns a single student
      operationId: getStudentById
      parameters:
        - name: studentId
          in: path
          description: ID of student to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentDTO'
        "400":
          description: Invalid ID supplied
          content: {}
        "404":
          description: StudentDTO not found
          content: {}
    put:
      tags:
        - student
      summary: Update an existing student
      description: This can only be done by the admin.
      operationId: updateStudent
      parameters:
        - name: studentId
          in: path
          description: ID of student that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Updated student object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentDTO'
        required: true
      responses:
        "200":
          description: StudentDTO updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentDTO'
        "400":
          description: Invalid ID supplied
          content: {}
        "404":
          description: StudentDTO not found
          content: {}
      x-codegen-request-body-name: body
    delete:
      tags:
        - student
      summary: Delete student
      description: This can only be done by the admin.
      operationId: deleteStudent
      parameters:
        - name: studentId
          in: path
          description: ID of student that needs to be deleted
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: StudentDTO deleted successfully
          content: {}
        "400":
          description: Invalid ID supplied
          content: {}
        "404":
          description: StudentDTO not found
          content: {}
  /tutor:
    get:
      tags:
        - tutor
      summary: Get all tutors
      description: Returns all tutors in the system
      operationId: getAllTutors
      responses:
        '200':
          description: Retrieved tutors successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TutorDTO'
        '400':
          description: Error retrieving tutors
          content: { }
    post:
      tags:
        - tutor
      summary: Add a new tutor
      description: This can only be done by the admin.
      operationId: addTutor
      requestBody:
        description: Tutor object that needs to be added to the system
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TutorDTO'
        required: true
      responses:
        "201":
          description: Tutor added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TutorDTO'
        "400":
          description: Invalid input
          content: {}
      x-codegen-request-body-name: body
  /tutor/{tutorId}:
    get:
      tags:
        - tutor
      summary: Get tutor by ID
      description: Returns a single tutor
      operationId: getTutorById
      parameters:
        - name: tutorId
          in: path
          description: ID of tutor to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TutorDTO'
        "400":
          description: Invalid ID supplied
          content: {}
        "404":
          description: Tutor not found
          content: {}
    put:
      tags:
        - tutor
      summary: Update an existing tutor
      description: This can only be done by the admin.
      operationId: updateTutor
      parameters:
        - name: tutorId
          in: path
          description: ID of tutor that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Updated tutor object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TutorDTO'
        required: true
      responses:
        "200":
          description: Tutor updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TutorDTO'
        "400":
          description: Invalid ID supplied
          content: {}
        "404":
          description: Tutor not found
          content: {}
      x-codegen-request-body-name: body
    delete:
      tags:
        - tutor
      summary: Delete tutor
      description: This can only be done by the admin.
      operationId: deleteTutor
      parameters:
        - name: tutorId
          in: path
          description: ID of tutor that needs to be deleted
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Tutor deleted successfully
          content: {}
        "400":
          description: Invalid ID supplied
          content: {}
        "404":
          description: Tutor not found
          content: {}
  /subject:
    get:
      tags:
        - subject
      summary: Get all subject
      description: Returns all subjects in the system
      operationId: getAllSubject
      responses:
        '200':
          description: Retrieved subjects successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubjectDTO'
        '400':
          description: Error retrieving subjects
          content: { }
    post:
      tags:
        - subject
      summary: Add a new subject
      description: This can only be done by the admin.
      operationId: addSubject
      requestBody:
        description: SubjectDTO object that needs to be added to the system
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectDTO'
        required: true
      responses:
        "201":
          description: SubjectDTO added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectDTO'
        "400":
          description: Invalid input
          content: {}
      x-codegen-request-body-name: body
  /subject/{subjectId}:
    get:
      tags:
        - subject
      summary: Get subject by ID
      description: Returns a single subject
      operationId: getSubjectById
      parameters:
        - name: subjectId
          in: path
          description: ID of subject to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectDTO'
        "400":
          description: Invalid ID supplied
          content: { }
        "404":
          description: Subject not found
          content: { }
    put:
      tags:
        - subject
      summary: Update an existing subject
      description: This can only be done by the admin.
      operationId: updateSubject
      parameters:
        - name: subjectId
          in: path
          description: ID of subject that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Updated subject object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectDTO'
        required: true
      responses:
        "200":
          description: SubjectDTO updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectDTO'
        "400":
          description: Invalid ID supplied
          content: {}
        "404":
          description: SubjectDTO not found
          content: {}
      x-codegen-request-body-name: body
    delete:
      tags:
        - subject
      summary: Delete subject
      description: This can only be done by the admin.
      operationId: deleteSubject
      parameters:
        - name: subjectId
          in: path
          description: ID of subject that needs to be deleted
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: SubjectDTO deleted successfully
          content: {}
        "400":
          description: Invalid ID supplied
          content: {}
        "404":
          description: SubjectDTO not found
          content: {}
...
components:
  schemas:
...
    StudentDTO:
      required:
        - email
        - name
        - lastName
        - password
        - phoneNumber
        - studentNumber
      type: object
      properties:
        studentId:
          type: integer
          format: int64
        name:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        password:
          type: string
        studentNumber:
          type: integer
          format: int64
    TutorDTO:
      required:
        - email
        - name
        - password
        - lastName
        - phoneNumber
      type: object
      properties:
        tutorId:
          type: integer
          format: int64
        name:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        password:
          type: string
    SubjectDTO:
      required:
        - subjectCode
        - subjectName
      type: object
      properties:
        subjectId:
          type: integer
          format: int64
        subjectCode:
          type: string
        subjectName:
          type: string
...